+++ src/com/fsck/k9/mail/Address.java
@@ -255,7 +255,7 @@ public class Address
         return toFriendly((Contacts)null, 0);
     }
 
-    public CharSequence toFriendly(Contacts contacts, long registered_name_color)
+    public CharSequence toFriendly(Contacts contacts, int registered_name_color)
     {
         if (contacts != null)
         {
@@ -264,7 +264,7 @@ public class Address
             if (name != null && name != NO_ENTRY)
             {
                 SpannableString sname = new SpannableString(name);
-                sname.setSpan(new ForegroundColorSpan((int)registered_name_color), // TODO: should not cast
+                sname.setSpan(new ForegroundColorSpan(registered_name_color),
                         0,
                         sname.length(),
                         Spannable.SPAN_EXCLUSIVE_EXCLUSIVE
@@ -285,7 +285,7 @@ public class Address
                             sContactsName.put(mAddress, name);
 
                             SpannableString sname = new SpannableString(name);
-                            sname.setSpan(new ForegroundColorSpan((int)registered_name_color), // TODO: should not cast
+                            sname.setSpan(new ForegroundColorSpan(registered_name_color),
                                     0,
                                     sname.length(),
                                     Spannable.SPAN_EXCLUSIVE_EXCLUSIVE
@@ -320,7 +320,7 @@ public class Address
         return toFriendly(addresses, null, 0);
     }
 
-    public static CharSequence toFriendly(Address[] addresses, Contacts contacts, long registered_name_color)
+    public static CharSequence toFriendly(Address[] addresses, Contacts contacts, int registered_name_color)
     {
         if (addresses == null)
         {
